Added in custom cursors
Added in custom cursors when hovering over certain cells. Added in the ability to set cells to white and erase marks by pressing down the shift and alt keys. Started the function to put in the peek functionality. Added in the function to show the solution. Prevented the default browser actions for when the cells are being painted.